requirements.txt
pandas==1.5.1
requests==2.28.1
pyyaml==6.0
pybit==2.4.1
matplotlib==3.7.0
mplfinance